+++ libs/angular-components/tedi/components/notifications/alert/alert.component.ts
@@ -9,8 +9,10 @@ import {
   AfterViewInit,
 } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { IconComponent } from "../../base/icon/icon.component";
-import { ClosingButtonComponent } from "../../../../community/components/buttons/closing-button/closing-button.component";
+import {
+  IconComponent,
+  ClosingButtonComponent,
+} from "@tehik-ee/tedi-angular/tedi";
 
 export type AlertRole = "alert" | "status" | "none";
 export type AlertType = "info" | "success" | "warning" | "danger";