+++ libs/angular-components/tedi/components/overlay/index.ts
@@ -1,1 +1,1 @@
-export * from "./tooltip/tooltip.component";
+export * from "./tooltip";